Move transform tests to tests/
